% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{cb_create}
\alias{cb_create}
\title{Generate a codebook object}
\usage{
cb_create(
  data,
  metadata = NULL,
  ...,
  .name = name,
  .var_label = label,
  .val_labels = val_labels,
  .user_missing = NULL,
  .val_labs_sep1 = NULL,
  .val_labs_sep2 = NULL,
  .rmv_html = !name,
  .rmv_line_breaks = !name,
  .separate_missings = c("if_any", "yes", "no"),
  .user_missing_conflict = c("metadata", "missing_label")
)
}
\arguments{
\item{data}{A data frame exported or retrieved from REDCap.}

\item{metadata}{A data frame containing metadata, such as variable labels and value
labels.}

\item{...}{Additional columns from \code{metadata} to preserve in the final codebook.
New names can be assigned by passing named arguments. Columns for variable
name, form, variable label, and value labels are included by default.}

\item{.name, .var_label, .val_labels}{Columns in \code{metadata} containing variable
name, variable label, and value labels, respectively. If \code{metadata} is provided,
\code{.name} must be specified. \code{.var_label} and \code{.val_labels} may be set to \code{NULL}
to omit.}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. See "Specifying user missing values" below for examples.}

\item{.val_labs_sep1, .val_labs_sep2}{Regex patterns separating value labels
in \code{metadata}. \code{.val_labs_sep1} separates values from labels, and \code{.val_labs_sep2}
separates value/label pairs. e.g., if value labels are in format \code{"1, First label|2, Second label"},
set \code{.val_labs_sep1} to \code{","} and \code{.val_labs_sep2} to \code{"\\\\|"}.}

\item{.rmv_html}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Codebook columns from which
HTML tags should be removed.}

\item{.rmv_line_breaks}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Codebook columns
from which line breaks should be removed.}

\item{.separate_missings}{Include value labels for user missing values in a separate
column? The default, \code{"if_any"}, adds the column only if user missings are
specified for at least one variable.}

\item{.user_missing_conflict}{If different labels for a value are provided in
metadata and user missings, which should be used?}
}
\value{
An \code{"li_codebook"} object, consisting of (1) a tibble summarizing the passed
dataset and (2) attributes containing the passed dataset (in several formats)
and additional metadata. Specifically:
\itemize{
\item A tibble with columns:
\itemize{
\item \code{name}: variable name
\item \code{type}: variable type
\item \code{label}: variable label
\item \code{value_labels}: value labels
\item \code{user_missing}: optional column, depending on value of \code{.separate_missings},
with value labels for user missing values
\item \code{missing}: proportion missing
\item additional columns if specified in \code{...}
}
\item Attributes:
\itemize{
\item Transformed versions of the passed dataset. See \code{\link[=cb_get_data]{cb_get_data()}}.
\item Lookup tables and other metadata used internally: \code{"user_missing"}, \code{"vals_by_label"},
\code{"labs_by_value"}, \code{"miss_propagate"}, \code{"factors"}, \code{"n_obs"}, \code{"n_vars"}
}
}
}
\description{
\code{cb_create()} builds an object of class \code{"li_codebook"} from a dataset and optional
metadata. The resulting object can be used to write an Excel workbook with variable
and data summaries (using \code{\link[=cb_write]{cb_write()}}), extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}),
or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}} and \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}}).
}
\section{Specifying user missing values}{

User missing values are defined by passing a formula or list of formulas to the
\code{.user_missing} argument. Formulas should specify variables on the left-hand
side and user missing values for those variables on the right-hand side:
\preformatted{
cb <- cb_create_redcap(data, metadata, .user_missing = var1 ~ 99)
}
The same user missings can be applied to multiple variables using
\link[dplyr:dplyr_tidy_select]{tidyselect} expressions:
\preformatted{
# for variables `var1` through `var5`
.user_missing = var1:var5 ~ 99

# for all numeric variables, plus `var6` and `var7`
.user_missing = c(where(is.numeric), var6, var7) ~ c(-9, -8, -7)
}
Different user missings can be applied to different variables using a list of
formulas:
\preformatted{
.user_missing = list(
  starts_with("status") ~ c(98, 99),
  var7:var10 ~ 97
)
}
User missing values may optionally be named to set value labels:
\preformatted{
.user_missing = where(is.numeric) ~
  c(Declined = -97, "Don't know" = -98, "Not applicable" = -99)
}
If labels set in \code{.user_missing} conflict with those in \code{metadata}, \code{.user_missing_conflict}
controls which labels are used.
}

