% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create.r, R/cb_create_redcap.r
\name{cb_create_options}
\alias{cb_create_options}
\alias{cb_create_redcap_options}
\title{Additional options for codebook creation}
\usage{
cb_create_options(
  ...,
  include_types = TRUE,
  include_r_classes = FALSE,
  rmv_html = TRUE,
  rmv_line_breaks = TRUE,
  user_missing_col = c("if_any", "yes", "no"),
  user_missing_conflict = c("metadata", "missing_label"),
  user_missing_incompatible = c("ignore", "warn", "error")
)

cb_create_redcap_options(
  ...,
  include_types = TRUE,
  include_r_classes = FALSE,
  rmv_html = TRUE,
  rmv_line_breaks = TRUE,
  user_missing_col = c("if_any", "yes", "no"),
  user_missing_conflict = c("metadata", "missing_label"),
  user_missing_incompatible = c("ignore", "warn", "error"),
  name = field_name,
  var_label = field_label,
  val_labels = select_choices_or_calculations,
  type = field_type,
  form = form_name,
  val_labs_sep1 = ", ",
  val_labs_sep2 = "\\\\|",
  coerce_integers = TRUE,
  checkbox_resp_values = FALSE,
  propagate_checkbox_missings = TRUE
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{include_types}{Include a column listing simplified type for each variable?
(e.g,. \code{"categorical"}, \code{"date-time"}.)}

\item{include_r_classes}{Include a column listing class(es) of each variable?
(e.g., \code{"factor"}, \code{"POSIXct, POSIXt"}.)}

\item{rmv_html}{Should HTML tags be removed from metadata (e.g., from variable
and value labels)?}

\item{rmv_line_breaks}{Should line breaks be removed from metadata (e.g., from
variable and value labels)? If \code{TRUE}, line breaks will be replaced with \code{" / "}.}

\item{user_missing_col}{Include value labels for user missing values in a separate
column? The default, \code{"if_any"}, adds the column only if user missings are
specified for at least one variable.}

\item{user_missing_conflict}{If labels passed to \code{.user_missing} conflict with
value labels in metadata, which should be used?}

\item{user_missing_incompatible}{How to handle variables specified in \code{.user_missing}
that aren't compatible with user missing values (e.g., logical, Date, or POSIXt)?}

\item{name, var_label, val_labels, type}{For REDCap data, columns in \code{metadata} containing variable
name, variable label, value labels, and variable type, respectively.}

\item{form}{For REDCap data, column in \code{metadata} containing form names. (Set to \code{NULL} to omit.)}

\item{val_labs_sep1, val_labs_sep2}{For REDCap data, regex patterns separating value labels
in \code{metadata}. \code{val_labs_sep1} separates values from labels, and \code{val_labs_sep2}
separates value/label pairs from one another. e.g., if value labels are in
the format \code{"1, First label|2, Second label"}, set \code{val_labs_sep1} to \code{","}
and \code{val_labs_sep2} to \code{"\\\\|"}.}

\item{coerce_integers}{For REDCap data, should variables listed as "integer" in \code{metedata$text_validation_type_or_show_slider_number}
be coerced to integer?}

\item{checkbox_resp_values}{For REDCap data, should checkbox values use labels in \code{metadata} (\code{TRUE})
or "Yes" / "No" (\code{FALSE})? See "Checkbox data handling"  on the \code{cb_create_redcap()}
help page.}

\item{propagate_checkbox_missings}{For REDCap data, should user missing values in a checkbox group
be propagated across all variables in the group? See "Checkbox data handling"
on the \code{cb_create_redcap()} help page.}
}
\description{
Additional options for use by \code{cb_create()}.
}
