% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{cb_create_rc}
\alias{cb_create_rc}
\title{Generate a codebook object from REDCap data}
\usage{
cb_create_rc(
  data,
  metadata,
  ...,
  .name = field_name,
  .var_label = field_label,
  .val_labels = select_choices_or_calculations,
  .form = form_name,
  .user_missing = NULL,
  .val_labs_sep1 = ", ",
  .val_labs_sep2 = "\\\\|",
  .rmv_html = !name,
  .rmv_line_breaks = !name,
  .coerce_integers = TRUE,
  .checkbox_resp_values = FALSE,
  .propagate_checkbox_missings = TRUE,
  .separate_missings = c("if_any", "yes", "no"),
  .user_missing_conflict = c("metadata", "missing_label")
)
}
\arguments{
\item{data}{A data frame exported or retrieved from REDCap.}

\item{metadata}{A data frame containing the REDCap codebook associated with \code{data}.}

\item{...}{Additional columns from \code{metadata} to preserve in the final codebook.
New names can be assigned by passing named arguments. Columns for variable
name, form, variable label, and value labels are included by default.}

\item{.name, .var_label, .val_labels}{Columns in \code{metadata} containing variable
name, variable label, and value labels, respectively.}

\item{.form}{Column in \code{metadata} containing form names. (Set to \code{NULL} to omit.)}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. See "Specifying user missing values" below for examples.}

\item{.val_labs_sep1, .val_labs_sep2}{Regex patterns separating value labels
in \code{metadata}. e.g., if value labels are in format \code{"1, First label|2, Second label"},
set \code{.val_labs_sep1} to \code{","} and \code{.val_labs_sep2} to \code{"\\\\|"}.}

\item{.rmv_html}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Codebook columns from which
HTML tags should be removed.}

\item{.rmv_line_breaks}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Codebook columns
from which line breaks should be removed.}

\item{.coerce_integers}{Should variables listed as "integer" in \code{metedata$text_validation_type_or_show_slider_number}
be coerced to integer?}

\item{.checkbox_resp_values}{Should checkbox values use labels in \code{metadata}
(\code{TRUE}) or "Yes" / "No" (\code{FALSE})? See "Checkbox data handling" below.}

\item{.propagate_checkbox_missings}{Should user missing values in a checkbox
group be propagated across all variables in the group? See "Checkbox data handling"
below.}

\item{.separate_missings}{Include value labels for user missing values in a separate
column? The default, \code{"if_any"}, adds the column only if user missings are
specified for at least one variable.}

\item{.user_missing_conflict}{If different labels for a value are provided in
metadata and user missings, which should be used?}
}
\value{
An \code{"li_codebook"} object, consisting of (1) a tibble summarizing the passed
dataset and (2) attributes containing the passed dataset (in several formats)
and additional metadata. Specifically:
\itemize{
\item A tibble with columns:
\itemize{
\item \code{name}: variable name
\item \code{form}: form name
\item \code{type}: variable type
\item \code{label}: variable label
\item \code{value_labels}: value labels
\item \code{user_missing}: optional column, depending on value of \code{.separate_missings},
with value labels for user missing values
\item \code{missing}: proportion missing
\item additional columns if specified in \code{...}
}
\item Attributes:
\itemize{
\item Transformed versions of the passed dataset. See \code{\link[=cb_get_data]{cb_get_data()}}.
\item Lookup tables and other metadata used internally: \code{"user_missing"}, \code{"vals_by_label"},
\code{"labs_by_value"}, \code{"miss_propagate"}, \code{"factors"}, \code{"n_obs"}, \code{"n_vars"}
}
}
}
\description{
\code{cb_create_rc()} builds an object of class \code{"li_codebook"} from a dataset and
corresponding codebook exported from REDCap. The resulting object can be used
to write an Excel workbook with variable and data summaries (using \code{\link[=cb_write]{cb_write()}}),
extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}), or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}}
and \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}}).

This variant of \code{cb_create()} includes functionality specific to REDCap data
and metadata, including:
\itemize{
\item Defaults for typical REDCap metadata column names
\item Includes form field by default
\item Unpacking, labelling, and optional missing propagation for checkbox data
\item Optional coercion for character variables marked as "integer" in \code{metedata$text_validation_type_or_show_slider_number}
}
}
\section{Specifying user missing values}{

User missing values are defined by passing a formula or list of formulas to the
\code{.user_missing} argument. Formulas should specify variables on the left-hand
side and user missing values for those variables on the right-hand side:
\preformatted{
cb <- cb_create_rc(data, metadata, .user_missing = var1 ~ 99)
}
The same user missings can be applied to multiple variables using
\link[dplyr:dplyr_tidy_select]{tidyselect} expressions:
\preformatted{
# for variables `var1` through `var5`
.user_missing = var1:var5 ~ 99

# for all numeric variables, plus `var6` and `var7`
.user_missing = c(where(is.numeric), var6, var7) ~ c(-9, -8, -7)

# for all variables
.user_missing = everything() ~ c(-9, -8)
}
Different user missings can be applied to different variables using a list of
formulas:
\preformatted{
.user_missing = list(
  starts_with("status") ~ c(98, 99),
  var7:var10 ~ 97
)
}
User missing values may optionally be named to set value labels:
\preformatted{
.user_missing = where(is.numeric) ~
  c(Declined = -97, "Don't know" = -98, "Not applicable" = -99)
}
If labels set in \code{.user_missing} conflict with those in \code{metadata}, \code{.user_missing_conflict}
controls which labels are used.
}

\section{Checkbox data handling}{

\subsection{Value labels}{

Data from REDCap checkboxes yields one variable in the dataset for each response
option. These will be labelled generically with \code{"Yes"} or \code{"No"}, unless \code{.checkbox_resp_values}
is \code{TRUE}, in which case response-specific labels from \code{metadata} will be used.
For example, if a checkbox group has options "In the past year," "More than a
year ago," and "Never," corresponding to variables \code{chk_var1___0}, \code{chk_var1___1},
and \code{chk_var1___2}: if \code{.checkbox_resp_values} is \code{FALSE}, all of these will
have values:
\itemize{
\item \code{chk_var1___0}, \code{chk_var1___1}, \code{chk_var1___2}: 0 = "No"; 1 =  "Yes".
}

If \code{.checkbox_resp_values} is \code{TRUE}, each variable will have unique labels:
\itemize{
\item \code{chk_var1___0}: 0 = "Not selected," 1 = "In the past year"
\item \code{chk_var1___1}: 0 = "Not selected," 1 = "More than a year ago"
\item \code{chk_var1___2}: 0 = "Not selected," 0 = "Never"
}
}

\subsection{Missing value propagation}{

If \code{.propagate_checkbox_missings} is \code{TRUE}, missing values in a checkbox group
variable will be propagated to all variables in the group. For example, given
a checkbox group with options "Pregnant," "Not pregnant," and "Not applicable,"
corresponding to variables \code{chk_preg_0___0}, \code{chk_preg_0___1}, and \code{chk_preg_0____9},
and assuming that \code{-9} is specified as a user missing value. If \code{.propagate_checkbox_missings}
is \code{TRUE}, \code{chk_preg_0___0} and \code{chk_preg_0___1} will be set to \code{-9} if \code{chk_preg_0____9}
is \code{1}. Otherwise, these columns will remain as \code{0} where \code{chk_preg_0____9} is \code{1}.
}
}

