% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_redcap.r
\name{cb_create_redcap}
\alias{cb_create_redcap}
\title{Generate a codebook object from REDCap data}
\usage{
cb_create_redcap(
  data,
  metadata,
  ...,
  .user_missing = NULL,
  .split_var_labels = NULL,
  .options = cb_create_redcap_options()
)
}
\arguments{
\item{data}{A data frame exported or retrieved from REDCap.}

\item{metadata}{A data frame containing the REDCap codebook associated with \code{data}.}

\item{...}{Additional columns from \code{metadata} to preserve in the final codebook.
New names can be assigned by passing named arguments. Columns for variable
name, form, variable label, and value labels are included by default.}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. If left-hand side is omitted, defaults to \code{tidyselect::everything()}.
See "Specifying user missing values" in \code{\link[=cb_create]{cb_create()}} documentation  for examples.}

\item{.split_var_labels}{A \code{\link[dplyr:dplyr_tidy_select]{tidyselect}} expression or list of tidyselect
expressions, indicating (sets of) variable labels with a common stem that should
be extracted into a separate column.}

\item{.options}{Additional options to use for codebook creation. Must be the result
from a call to \code{cb_create_redcap_options()} or \code{cb_create_options()}. See \code{?cb_create_redcap_options}
for available options.}
}
\value{
An \code{"li_codebook"} object, consisting of a tibble summarizing the passed
dataset and attributes containing additional metadata. The tibble includes columns:
\itemize{
\item \code{name}: variable name
\item \code{form}: form name
\item \code{type}: column containing simplified variable type
\item \code{class}: optional column containing class(es) of each variable
\item \code{label_stem}: optional column containing variable label stems, if any variables
are specified in \code{.split_var_labels}
\item \code{label}: variable label
\item \code{values}: values, with labels if applicable
\item \code{user_missing}: optional column showing user missing values, with labels
if applicable. By default, this column is included only if user missings
are specified for at least one variable. This behavior can be changed using
the \code{user_missing_col} argument to \code{cb_create_options()}.
\item \code{missing}: proportion missing
\item additional columns if specified in \code{...}
}
}
\description{
\code{cb_create_redcap()} builds an object of class \code{"li_codebook"} from a dataset and
corresponding codebook exported from REDCap. The resulting object can be used
to write an Excel workbook with variable and data summaries (using \code{\link[=cb_write]{cb_write()}}),
extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}), or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}},
\code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}}, \code{\link[=cb_summarize_text]{cb_summarize_text()}}).

This variant of \code{\link[=cb_create]{cb_create()}} includes functionality specific to REDCap data
and metadata, including:
\itemize{
\item Defaults for typical REDCap metadata column names
\item Includes form field by default
\item Unpacking, labelling, and optional missing propagation for checkbox data
\item Optional coercion for character variables marked as "integer" in \code{metedata$text_validation_type_or_show_slider_number}
}

All of these behaviors can be controlled using the \code{.options} argument.
}
\section{Checkbox data handling}{

\subsection{Value labels}{

Data from REDCap checkboxes yields one variable in the dataset for each response
option. By default, these will be labelled generically with \code{"Yes"} or \code{"No"}.
For example, consider a checkbox group with options "In the past year," "More than a
year ago," and "Never," corresponding to variables \code{chk_var1___0}, \code{chk_var1___1},
and \code{chk_var1___2}. By default, all of these will be given the same value labels:
\itemize{
\item \code{chk_var1___0}, \code{chk_var1___1}, \code{chk_var1___2}: 0 = "No"; 1 =  "Yes".
This behavior can be changed by setting \code{checkbox_resp_values = TRUE} in \code{cb_create_options()}.
In this case, response-specific labels from \code{metadata} will be used, so that
each variable will have unique labels:
\item \code{chk_var1___0}: 0 = "Not selected," 1 = "In the past year"
\item \code{chk_var1___1}: 0 = "Not selected," 1 = "More than a year ago"
\item \code{chk_var1___2}: 0 = "Not selected," 0 = "Never"
}
}

\subsection{Missing value propagation}{

By default, missing values in a checkbox group will be propagated to all variables
in the group. For example, consider a checkbox group with options "Pregnant,"
"Not pregnant," and "Not applicable," corresponding to variables \code{chk_preg_0___0},
\code{chk_preg_0___1}, and \code{chk_preg_0____9}, and assuming that \code{-9} is specified
as a user missing value. By default, \code{chk_preg_0___0} and \code{chk_preg_0___1} will
be set to \code{-9} if \code{chk_preg_0____9} is \code{1}. This behavior can be overridden by
setting \code{propagate_checkbox_missings = FALSE} in \code{cb_create_options()}, in which
case no values will be changed.
}
}

