% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_spss.r
\name{cb_create_spss}
\alias{cb_create_spss}
\title{Generate a codebook object from an SPSS dataset}
\usage{
cb_create_spss(
  data,
  .user_missing = NULL,
  .split_var_labels = NULL,
  .options = cb_create_options()
)
}
\arguments{
\item{data}{A data frame imported from SPSS using (imported using \code{haven::read_spss()},
\code{read_sav()}, or \code{read_por()}).}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. If left-hand side is omitted, defaults to \code{tidyselect::everything()}.
See "Specifying user missing values" in \code{\link[=cb_create]{cb_create()}} documentation for examples.}

\item{.split_var_labels}{A \code{\link[dplyr:dplyr_tidy_select]{tidyselect}} expression or list of tidyselect
expressions, indicating (sets of) variable labels with a common stem that should
be extracted into a separate column.}

\item{.options}{Additional options to use for codebook creation. Must be the result
from a call to \code{cb_create_options()}. See that function's help page for available
options.}
}
\value{
An \code{"li_codebook"} object, consisting of a tibble summarizing the passed
dataset and attributes containing additional metadata. The tibble includes columns:
\itemize{
\item \code{name}: variable name
\item \code{type}: column containing simplified variable type
\item \code{class}: optional column containing class(es) of each variable
\item \code{label_stem}: optional column containing variable label stems, if any variables
are specified in \code{.split_var_labels}
\item \code{label}: variable label
\item \code{values}: values, with labels if applicable
\item \code{user_missing}: optional column showing user missing values, with labels
if applicable. By default, this column is included only if user missings
are specified for at least one variable. This behavior can be changed using
the \code{user_missing_col} argument to \code{cb_create_options()}.
\item \code{missing}: proportion missing
}
}
\description{
\code{cb_create_spss()} builds an object of class \code{"li_codebook"} from an imported
SPSS dataset. Metadata including variable labels, value labels, and user missing
values are extracted from the imported dataset. (User missing values can also
be set using the \code{.user_missing} argument.)`

The resulting object can be used to write an Excel workbook with variable and
data summaries (using \code{\link[=cb_write]{cb_write()}}), extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}),
or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}}, \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}},
\code{\link[=cb_summarize_text]{cb_summarize_text()}}).
}
