% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create.r
\name{cb_get_data}
\alias{cb_get_data}
\title{Extract data from a codebook object}
\usage{
cb_get_data(cb, format = c("factors", "haven"))
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{format}{Format of the returned data, either \code{"factors"} or \code{"haven"};
see below for details.}
}
\value{
A tibble with variables formatted based on the \code{format} argument.
\itemize{
\item For \code{"factors"}, all variables with value labels are converted to factors,
and all user missings are converted to \code{NA}.
\item For \code{"haven"}, variable labels, value labels, and user missings are encoded
using class \code{\link[haven:labelled]{"haven_labelled_spss"}}`.
}

Both formats may also reflect transformations made by variants of \code{\link[=cb_create]{cb_create()}}.
In particular, for codebooks created using \code{\link[=cb_create_redcap]{cb_create_redcap()}}, integer coercion
and propagation of user missings across checkbox variables.
}
\description{
Codebook objects created by \code{\link[=cb_create]{cb_create()}} and friends contain several transformed
versions of the originally passed dataset. These can be extracted using \code{cb_get_data()}.
}
