% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{cb_summarize_categorical}
\alias{cb_summarize_categorical}
\title{Summarize categorical variables from a codebook object}
\usage{
cb_summarize_categorical(
  cb,
  group_by = NULL,
  prefixed = TRUE,
  detail_missing = missing(group_by),
  detail_na_label = "NA"
)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{cb_create()} or
a variant.}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by.}

\item{prefixed}{Should value labels be prefixed with the corresponding value?
e.g., \code{TRUE} yields \code{"[1] Value One"}; \code{FALSE} yields \code{"Value One"}.}

\item{detail_missing}{Include detailed missing value information? Currently supported
only when no grouping variables are specified.}

\item{detail_na_label}{Label used for \code{NA} values when \code{detail_missing} is \code{TRUE}.}
}
\value{
A tibble with frequency information for each categorical variable.
}
\description{
\code{cb_summarize_categorical()} generates a frequencies table for all categorical
variables from a codebook object, optionally by group.
}
