% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summarize.r
\name{cb_summarize_character}
\alias{cb_summarize_character}
\title{Summarize string variables from a codebook object}
\usage{
cb_summarize_character(
  cb,
  n_char_vals = 5,
  detail_missing = TRUE,
  detail_na_label = "NA"
)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{n_char_vals}{Frequencies for the \code{n_char_vals} most frequent values will
be included.}

\item{detail_missing}{Include detailed missing value information?}

\item{detail_na_label}{Label used for \code{NA} values when \code{detail_missing} is \code{TRUE}.}
}
\value{
A tibble with columns:
\itemize{
\item \code{name}: variable name
\item \code{label_stem}: optional column containing variable label stems; included if
\code{cb} includes a \code{label_stem} column and at least one character variable
has a non-missing label stem.
\item \code{label}: variable label
\item \code{n_unique}: number of unique non-missing values
\item \code{is_missing}: optional column indicating if \code{value} is a missing value. Included
if \code{detail_missing} is \code{TRUE}.
\item \code{value}: the most prevalent unique values for the variable. If there are
more than \code{n_char_vals} + 1 unique values, the \code{n_char_vals} most common
non-missing values will be included. (All missing values will always be included.)
\item \code{n}: number of observations
\item \code{pct_of_all}: proportion of all (non-missing and missing) observations
\item \code{pct_of_valid}: for non-missing values, proportion of all non-missing observations
\item \code{pct_of_missing}: optional column showing, for missing values, proportion
of all missing observations. Included if \code{detail_missing} is \code{TRUE}.
}
}
\description{
\code{cb_summarize_character()} generates a summary table for all character variables
from a codebook object, including number of unique values, frequencies for the
most common values, and missing value information. Note that character variables
of class \code{"haven_labelled"} are treated as categorical; see \code{cb_summarize_categorical()}.
}
