% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighthouse.codebook-package.R
\docType{package}
\name{lighthouse.codebook-package}
\alias{lighthouse.codebook}
\alias{lighthouse.codebook-package}
\title{lighthouse.codebook: Summarize Datasets for Lighthouse Institute Projects}
\description{
Functions to generate codebooks and dataset summaries for Lighthouse Institute projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ccsarapas/lighthouse.codebook}
  \item Report bugs at \url{https://github.com/ccsarapas/lighthouse.codebook/issues}
}

}
\author{
\strong{Maintainer}: Casey Sarapas \email{ccsarapas@chestnut.org} (\href{https://orcid.org/0000-0002-8668-1462}{ORCID})

}
\keyword{internal}
