% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{stats}
\alias{stats}
\alias{skew}
\alias{kurtosis}
\alias{spread}
\alias{min_if_any}
\alias{max_if_any}
\alias{se_mean}
\title{Statistics for numeric summaries}
\usage{
skew(x, adjusted = TRUE, na.rm = FALSE)

kurtosis(x, adjusted = TRUE, excess = TRUE, na.rm = FALSE)

spread(x, na.rm = FALSE)

min_if_any(..., na.rm = TRUE)

max_if_any(..., na.rm = TRUE)

se_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{adjusted}{If \code{TRUE}, returns adjusted skewness (\emph{G}\if{html}{\out{<sub>}}1\if{html}{\out{</sub>}}) or kurtosis
(\emph{G}\if{html}{\out{<sub>}}2\if{html}{\out{</sub>}}) by applying a small-sample correction. If \code{FALSE}, returns
the unadjusted skewness (\emph{g}\if{html}{\out{<sub>}}1\if{html}{\out{</sub>}}) or kurtosis (\emph{g}\if{html}{\out{<sub>}}2\if{html}{\out{</sub>}}). (Note
that \code{TRUE} corresponds to behavior of software such as SPSS, SAS, and Excel.)}

\item{na.rm}{Should missing values be removed? (Note that \code{cb_summarize_numeric()}
automatically sets na.rm to \code{TRUE} for all functions).}

\item{excess}{If \code{TRUE}, returns excess kurtosis by (total kurtosis - 3).}

\item{...}{A numeric vector or vectors.}
}
\description{
Functions for computing summary statistics for use in \code{cb_summarize_numeric()}.
\itemize{
\item \code{skew()} and \code{kurtosis()} return adjusted skewness and kurtosis. (Unadjusted
estimates can be obtained by setting \code{adjusted = FALSE}.)
\item \code{spread()} returns the difference between a vector's minimum and maximum values.
\item \code{min_if_any()} and \code{max_if_any()} return minima and maxima with alternate behavior
if all values are missing. (Re-exported from the lighthouse package. See \code{lighthouse::min_if_any}
for more details.)
\item \code{se_mean()} returns the standard error of the mean. (Re-exported from the lighthouse
package. See \code{lighthouse::se_mean} for more details.)
}
}
