% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create.r
\name{cb_create}
\alias{cb_create}
\title{Generate a codebook object}
\usage{
cb_create(
  data,
  metadata = NULL,
  ...,
  .name = name,
  .var_label = label,
  .val_labels = val_labels,
  .user_missing = NULL,
  .split_var_labels = NULL,
  .include_types = !.include_r_classes,
  .include_r_classes = FALSE,
  .val_labs_sep1 = NULL,
  .val_labs_sep2 = NULL,
  .rmv_html = TRUE,
  .rmv_line_breaks = TRUE,
  .user_missing_col = c("if_any", "yes", "no"),
  .user_missing_conflict = c("metadata", "missing_label"),
  .user_missing_incompatible = c("ignore", "warn", "error")
)
}
\arguments{
\item{data}{A data frame.}

\item{metadata}{A data frame containing metadata, such as variable labels and value
labels.}

\item{...}{Additional columns from \code{metadata} to preserve in the final codebook.
New names can be assigned by passing named arguments. Columns for variable
name, form, variable label, and value labels are included by default.}

\item{.name, .var_label, .val_labels}{Columns in \code{metadata} containing variable
name, variable label, and value labels, respectively. If \code{metadata} is provided,
\code{.name} must be specified. \code{.var_label} and \code{.val_labels} may be set to \code{NULL}
to omit.}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. If left-hand side is omitted, defaults to \code{tidyselect::everything()}.
See "Specifying user missing values" below for examples.}

\item{.split_var_labels}{A \code{\link[dplyr:dplyr_tidy_select]{tidyselect}} expression or list of tidyselect
expressions, indicating (sets of) variable labels with a common stem that should
be extracted into a separate column.}

\item{.include_types}{Include a column listing simplified type for each variable?
(e.g,. \code{"categorical"}, \code{"date-time"}.)}

\item{.include_r_classes}{Include a column listing class(es) of each variable?
(e.g., \code{"factor"}, \code{"POSIXct, POSIXt"}.)}

\item{.val_labs_sep1, .val_labs_sep2}{Regex patterns separating value labels
in \code{metadata}. \code{.val_labs_sep1} separates values from labels, and \code{.val_labs_sep2}
separates value/label pairs. e.g., if value labels are in format \code{"1, First label|2, Second label"},
set \code{.val_labs_sep1} to \code{","} and \code{.val_labs_sep2} to \code{"\\\\|"}.}

\item{.rmv_html}{Should HTML tags be removed from metadata (e.g., from variable
and value labels)?}

\item{.rmv_line_breaks}{Should line breaks be removed from metadata (e.g., from
variable and value labels)? If \code{TRUE}, line breaks will be replaced with \code{" / "}.}

\item{.user_missing_col}{Include value labels for user missing values in a separate
column? The default, \code{"if_any"}, adds the column only if user missings are
specified for at least one variable.}

\item{.user_missing_conflict}{If different labels for a value are provided in
metadata and user missings, which should be used?}

\item{.user_missing_incompatible}{How to handle variables specified in \code{.user_missing}
that aren't compatible with user missing values (e.g., logical, Date, or POSIXt)?}
}
\value{
An \code{"li_codebook"} object, consisting of (1) a tibble summarizing the passed
dataset and (2) attributes containing the passed dataset (in several formats)
and additional metadata. Specifically:
\itemize{
\item A tibble with columns:
\itemize{
\item \code{name}: variable name
\item \code{type}: optional column containing simplified variable type
\item \code{class}: optional column containing class(es) of each variable
\item \code{label_stem}: optional column containing variable label stems, if any variables
are specified in \code{.split_var_labels}
\item \code{label}: variable label
\item \code{values}: values, with labels if applicable
\item \code{user_missing}: optional column, depending on value of \code{.user_missing_col},
showing user missing values, with labels if applicable
\item \code{missing}: proportion missing
\item additional columns if specified in \code{...}
}
\item Attributes:
\itemize{
\item Transformed versions of the passed dataset. See \code{\link[=cb_get_data]{cb_get_data()}}
\item Lookup tables and other metadata used internally.
}
}
}
\description{
\code{cb_create()} builds an object of class \code{"li_codebook"} from a dataset and optional
metadata. The resulting object can be used to write an Excel workbook with variable
and data summaries (using \code{\link[=cb_write]{cb_write()}}), extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}),
or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}}, \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}},
\code{\link[=cb_summarize_text]{cb_summarize_text()}}).
}
\section{Specifying user missing values}{

User missing values are defined by passing a formula or list of formulas to the
\code{.user_missing} argument. Formulas should specify variables on the left-hand
side and user missing values for those variables on the right-hand side:
\preformatted{
cb <- cb_create(data, metadata, .user_missing = var1 ~ 99)
}
The same user missings can be applied to multiple variables using \link[dplyr:dplyr_tidy_select]{tidyselect}
expressions.
\preformatted{
# for variables `var1` through `var5`
.user_missing = var1:var5 ~ 99

# for all numeric variables, plus `var6` and `var7`
.user_missing = c(where(is.numeric), var6, var7) ~ c(-9, -8, -7)

# omitted left-hand side defaults to `tidyselect::everything()`
.user_missing = ~ -99
}
Different user missings can be applied to different variables using a list of
formulas:
\preformatted{
.user_missing = list(
  starts_with("status") ~ c(98, 99),
  var7:var10 ~ 97
)
}
User missing values may optionally be named to set value labels:
\preformatted{
.user_missing = ~ c(Declined = -98, "Not applicable" = -99)
}
If labels set in \code{.user_missing} conflict with those in \code{metadata}, \code{.user_missing_conflict}
controls which labels are used.

User missings may be set for numeric, character, factor/ordered factor, and haven_labelled/haven_labelled_spss
vectors. For factors, user missings are set based on factor labels (not the underlying
integer codes). For \code{"haven_labelled"} vectors, user missings are set based on
values (not value labels). By default, variables with incompatible classes (e.g.,
logical, Date, POSIXt) will be ignored if specified in \code{.user_missing}. This
behavior can be changed using the \code{.user_missing_incompatible} argument.
}

\examples{
diamonds2 <- ggplot2::diamonds |>
  transform(
    carat_group = as.integer(cut(carat, breaks = 3, labels = 1:3)),
    price_group = as.integer(cut(
      price,
      breaks = c(0, 500, 1000, 2000, 5000, 10000, Inf),
      labels = 1:6,
      right = FALSE
    ))
  )

# basic codebook
cb_create(diamonds2)

# convert variables to factor to treat as categorical
diamonds2 |>
  transform(
    carat_group = factor(carat_group),
    price_group = factor(price_group)
  ) |>
  cb_create()

# provide metadata for variable and value labels
diamonds_meta <- data.frame(
  name = names(diamonds2),
  label = c(
    # from ?ggplot2::diamonds
    "price in US dollars ($326–$18,823)",
    "weight of the diamond (0.2–5.01)",
    "quality of the cut (Fair, Good, Very Good, Premium, Ideal)",
    "diamond colour, from D (best) to J (worst)",
    "a measurement of how clear the diamond is (I1 (worst), SI2, SI1, VS2, VS1, VVS2, VVS1, IF (best))",
    "length in mm (0–10.74)",
    "width in mm (0–58.9)",
    "depth in mm (0–31.8)",
    "total depth percentage = z / mean(x, y) = 2 * z / (x + y) (43–79)",
    "width of top of diamond relative to widest point (43–95)",
    "diamond carat (3 groups)",
    "diamond price (6 groups)"
  ),
  val_labels = c(
    rep(NA, 10),
    "1 = small; 2 = medium; 3 = large",
    "1 = <$500; 2 = $500-$999; 3 = $1,000-$1,999; 4 = $2,000-$4,999; 5 = $5,000-$9,999; 6 = $10,000+"
  )
)

cb_create(
  diamonds2, diamonds_meta,
  .val_labs_sep1 = " = ", .val_labs_sep2 = "; "
)
}
