% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_spss.r
\name{cb_create_spss}
\alias{cb_create_spss}
\title{Generate a codebook object from an SPSS dataset}
\usage{
cb_create_spss(
  data,
  .user_missing = NULL,
  .split_var_labels = NULL,
  .user_missing_col = c("if_any", "yes", "no"),
  .user_missing_conflict = c("val_label", "missing_label"),
  .user_missing_incompatible = c("ignore", "warn", "error")
)
}
\arguments{
\item{data}{A data frame exported or retrieved from REDCap.}

\item{.user_missing}{A formula or list of formulas specifying user missing values.
Formulas should specify variables on the left-hand side (as variable names
or \link[dplyr:dplyr_tidy_select]{tidyselect} expressions), and missing values on the
right-hand side. If left-hand side is omitted, defaults to \code{tidyselect::everything()}.
See "Specifying user missing values" in \code{\link[=cb_create]{cb_create()}} documentation for examples.}

\item{.split_var_labels}{A \code{\link[dplyr:dplyr_tidy_select]{tidyselect}} expression or list of tidyselect
expressions, indicating (sets of) variable labels with a common stem that should
be extracted into a separate column.}

\item{.user_missing_col}{Include value labels for user missing values in a separate
column? The default, \code{"if_any"}, adds the column only if user missings are
specified for at least one variable.}

\item{.user_missing_conflict}{If labels passed to \code{.user_missing} conflicts with
a value label in the \code{data}, which should be used?}

\item{.user_missing_incompatible}{How to handle variables specified in \code{.user_missing}
that aren't compatible with user missing values (e.g., logical, Date, or POSIXt)?}
}
\value{
An \code{"li_codebook"} object, consisting of (1) a tibble summarizing the passed
dataset and (2) attributes containing the passed dataset (in several formats)
and additional metadata. Specifically:
\itemize{
\item A tibble with columns:
\itemize{
\item \code{name}: variable name
\item \code{type}: variable type
\item \code{label_stem}: optional column containing variable label stems, if any variables
are specified in \code{.split_var_labels}
\item \code{label}: variable label
\item \code{value_labels}: value labels
\item \code{user_missing}: optional column, depending on value of \code{.user_missing_col},
with value labels for user missing values
\item \code{missing}: proportion missing
}
\item Attributes:
\itemize{
\item Transformed versions of the passed dataset. See \code{\link[=cb_get_data]{cb_get_data()}}.
\item Lookup tables and other metadata used internally.
}
}
}
\description{
\code{cb_create_spss()} builds an object of class \code{"li_codebook"} from an SPSS dataset
(imported using \code{haven::read_spss()}, \code{read_sav()}, or \code{read_por()}). Metadata
including variable labels, value labels, and user missing values are extracted
from the imported dataset. (User missing values can also be set using the \code{.user_missing}
argument.)`

The resulting object can be used to write an Excel workbook with variable and
data summaries (using \code{\link[=cb_write]{cb_write()}}), extract processed data (\code{\link[=cb_get_data]{cb_get_data()}}),
or generate dataset summaries (\code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}}, \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}},
\code{\link[=cb_summarize_text]{cb_summarize_text()}}).
}
