% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summarize.r
\name{cb_summarize_categorical}
\alias{cb_summarize_categorical}
\title{Summarize categorical variables from a codebook object}
\usage{
cb_summarize_categorical(
  cb,
  group_by = NULL,
  prefixed = TRUE,
  detail_missing = missing(group_by),
  detail_na_label = "NA"
)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by.}

\item{prefixed}{Should value labels be prefixed with the corresponding value?
e.g., \code{TRUE} yields \code{"[1] Value One"}; \code{FALSE} yields \code{"Value One"}.}

\item{detail_missing}{Include detailed missing value information? Currently supported
only when no grouping variables are specified.}

\item{detail_na_label}{Label used for \code{NA} values when \code{detail_missing} is \code{TRUE}.}
}
\value{
A tibble with columns:
\itemize{
\item optional grouping column(s) if specified in \code{group_by}
\item \code{name}: variable name
\item \code{label_stem}: optional column containing variable label stems; included if
\code{cb} includes a \code{label_stem} column and at least one categorical variable
has a non-missing label stem.
\item \code{label}: variable label
\item \code{is_missing}: optional column indicating if \code{value} is a missing value. Included
if \code{detail_missing} is \code{TRUE}.
\item \code{value}: variable value
\item \code{n}: number of observations
\item \code{pct_of_all}: proportion of all (non-missing and missing) observations
\item \code{pct_of_valid}: for non-missing values, proportion of all non-missing observations
\item \code{pct_of_missing}: optional column showing, for missing values, proportion
of all missing observations. Included if \code{detail_missing} is \code{TRUE}.
}
}
\description{
\code{cb_summarize_categorical()} generates a frequencies table for all categorical
variables from a codebook object, optionally by group. Variables with value labels,
factors (including ordered factors), and logical variables are treated as categorical.
}
