% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{cb_summarize_numeric}
\alias{cb_summarize_numeric}
\title{Summarize numeric variables from a codebook object}
\usage{
cb_summarize_numeric(cb, group_by = NULL)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{cb_create()} or
a variant.}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by.}
}
\value{
A tibble with summary statistics for each numeric variable.
}
\description{
\code{cb_summarize_numeric()} generates a summary table for all numeric variables
from a codebook object, optionally by group. Future releases will include options
to specify the summary statistics used. Currently, summary statistics are valid
n and \%; mean and SD; median, MAD, min, max, and range; skewness, and kurtosis.
}
