% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summarize.r
\name{cb_summarize_numeric}
\alias{cb_summarize_numeric}
\title{Summarize numeric variables from a codebook object}
\usage{
cb_summarize_numeric(cb, group_by = NULL, warn_if_none = TRUE)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by.}

\item{warn_if_none}{Should a warning be issued if there are no numeric variables
in \code{cb}?}
}
\value{
If there no numeric variables in \code{cb}, \code{NULL}. Otherwise, a tibble with
columns:
\itemize{
\item optional grouping column(s) if specified in \code{group_by}
\item \code{name}: variable name
\item \code{label_stem}: optional column containing variable label stems; included if
\code{cb} includes a \code{label_stem} column and at least one numeric variable has
a non-missing label stem.
\item \code{label}: variable label
\item \code{valid_n}, \code{valid_pct}: number and proportion of non-missing values
\item summary statistic columns: by default, these include \code{mean} and standard
deviation (\code{SD}); \code{median}, median absolute deviation (\code{MAD}), \code{min}, \code{max},
and \code{range}; skewness (\code{skew}), and kurtosis (\code{kurt}).
}
}
\description{
\code{cb_summarize_numeric()} generates a summary table for all numeric variables
from a codebook object, optionally by group. Future releases will include options
to specify the summary statistics used. Currently, summary statistics are valid
n and \%; mean and SD; median, MAD, min, max, and range; skewness, and kurtosis.
}
