% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_write.r
\name{cb_write}
\alias{cb_write}
\title{Write codebook and data summaries to an Excel workbook}
\usage{
cb_write(
  cb,
  file,
  dataset_name = NULL,
  incl_date = TRUE,
  incl_dims = TRUE,
  group_by = NULL,
  detail_missing = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{file}{Path to write to.}

\item{dataset_name}{Name of the dataset to display in workbook headers.}

\item{incl_date, incl_dims}{Should the date and/or dataset dimensions be included
in the Overview tab header?}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by. If specified, additional numeric and categorical summary tabs will be included
with decked heads for specified groups.}

\item{detail_missing}{Include detailed missing value information on categorical
summary tab?}

\item{overwrite}{Overwrite existing file?}
}
\value{
Invisibly returns the path to the written Excel file.
}
\description{
\code{cb_write()} writes an Excel workbook to disk with tabs including a codebook;
summary statistics for numeric variables; frequencies for categorical variables;
and optional grouped data summaries. For data summaires, variables with value
labels, factors (including ordered factors), and logical variables are treated
as categorical, while numeric and integer variables are treated as numeric.
}
