% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_write.r
\name{cb_write}
\alias{cb_write}
\title{Write codebook and data summaries to an Excel workbook}
\usage{
cb_write(
  cb,
  file,
  dataset_name = NULL,
  incl_date = TRUE,
  incl_dims = TRUE,
  hyperlinks = TRUE,
  group_by = NULL,
  detail_missing = c("if_any_user_missing", "yes", "no"),
  n_text_vals = 5,
  overwrite = TRUE
)
}
\arguments{
\item{cb}{An object of class \code{"li_codebook"} as produced by \code{\link[=cb_create]{cb_create()}} or
a variant.}

\item{file}{Path to write to.}

\item{dataset_name}{Name of the dataset to display in workbook headers.}

\item{incl_date, incl_dims}{Should the date and/or dataset dimensions be included
in the Overview tab header?}

\item{hyperlinks}{If \code{TRUE}, variable names on the Overview sheet will link
to corresponding rows on summary tabs and vice versa.}

\item{group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Column or columns to group
by. If specified, additional numeric and categorical summary tabs will be included
with decked heads for specified groups.}

\item{detail_missing}{Include detailed missing value information on categorical
and text summary tabs?}

\item{n_text_vals}{On the text summary tab, how many unique non-missing values
should be included for each variable? If there are more than \code{n_text_vals}
\itemize{
\item 1 unique values, the \code{n_text_vals} most common non-missing values will be included.
}}

\item{overwrite}{Overwrite existing file?}
}
\value{
Invisibly returns the path to the written Excel file. See []\code{cb_create()}]
and variants, \code{\link[=cb_summarize_numeric]{cb_summarize_numeric()}}, \code{\link[=cb_summarize_categorical]{cb_summarize_categorical()}}, and
\code{\link[=cb_summarize_text]{cb_summarize_text()}} for details on the objects written to the file.
}
\description{
\code{cb_write()} writes an Excel workbook to disk with tabs including a codebook;
summary statistics for numeric variables; frequencies for categorical variables;
truncated frequencies for text variables; and optional grouped summaries for numeric
and categorical variables. For data summaires, variables with value labels, factors,
and logical variables are treated as categorical, numeric and integer variables
are treated as numeric, and (unlabeled) character variables are treated as text.
Summary tabs will be omitted if there are no variables of the relevant type.
}
